god_bless_backend/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
psycopg2-binary
redis
django-redis
celery
channels
channels-redis
requests
pyfcm
django-cors-headers
daphne
djangorestframework-simplejwt
beautifulsoup4
Flask
phonenumbers
gunicorn
psutil
bleach>=6.0.0

# Production dependencies
gevent>=23.9.1
python-json-logger>=2.0.7
sentry-sdk>=1.40.0
django-celery-beat>=2.5.0